flask/flask_app/requirements.txt
gunicorn
configparser
pymongo
newspaper3k
flask
language-check
demjson
urllib3
